Deploy Lambda code through S3.
