+++ tests/utils/test_utility.py
@@ -9,19 +9,17 @@ class TestHelpers(unittest.TestCase):
     @mock.patch('boto3.client')
     def testAwsAccountId(self, patched_boto):
         """Tests the output of Helpers.aws_account_id."""
-        patched_boto.return_value.describe_security_groups.return_value = {
-          'SecurityGroups': [{ 'OwnerId': '123456654321' }]
-        }
+        patched_boto.return_value.get_caller_identity.return_value = {
+            'Arn': 'arn:aws:iam::123456654321:user/dliggat' }
 
         # Query for the account id; first to generate, second for a cache hit.
-        self.assertEqual(Helpers.aws_account_id(), 123456654321)
-        self.assertEqual(Helpers.aws_account_id(), 123456654321)
+        self.assertEqual(Helpers.aws_account_id(), '123456654321')
+        self.assertEqual(Helpers.aws_account_id(), '123456654321')
 
         # We should only ever the boto code once; the account value should
         # be memoized in the class after the initial invocation.
-        patched_boto.assert_called_once_with('ec2')
-        patched_boto.return_value.describe_security_groups.assert_called_once_with(
-            GroupNames=['default'])
+        patched_boto.assert_called_once_with('sts')
+        patched_boto.return_value.get_caller_identity.assert_called_once_with()
 
 
 def main():
